120 115 loose captures

+++ KioChess/Engine/Models/Config/MarginConfiguration.cs
@@ -3,7 +3,7 @@ namespace Engine.Models.Config;
 
 public class MarginConfiguration
 {
-    public short[][] AlphaMargins { get; set; }
-    public short[][] BetaMargins { get; set; }
-    public short[] DeltaMargins { get; set; }
+    public int[][] AlphaMargins { get; set; }
+    public int[][] BetaMargins { get; set; }
+    public int[] DeltaMargins { get; set; }
 }
